Moves SSH command running to vagrant.run_command
