+++ reprounzip-vagrant/reprounzip/unpackers/vagrant/__init__.py
@@ -17,7 +17,6 @@ import argparse
 import logging
 import os
 import paramiko
-from paramiko.client import MissingHostKeyPolicy
 import pickle
 from rpaths import PosixPath, Path
 import scp
@@ -31,18 +30,9 @@ from reprounzip.unpackers.common import COMPAT_OK, COMPAT_MAYBE, \
     composite_action, target_must_exist, make_unique_name, shell_escape, \
     select_installer, busybox_url, join_root, FileUploader, FileDownloader, \
     get_runs
-from reprounzip.unpackers.vagrant.interaction import interactive_shell
-from reprounzip.utils import irange, unicode_, iteritems, check_output
-
-
-class IgnoreMissingKey(MissingHostKeyPolicy):
-    """Policy that just ignores missing SSH host keys.
-
-    We are connecting to vagrant, checking the host doesn't make sense, and
-    accepting keys permanently is a security risk.
-    """
-    def missing_host_key(self, client, hostname, key):
-        pass
+from reprounzip.unpackers.vagrant.run_command import run_interactive, \
+    IgnoreMissingKey
+from reprounzip.utils import unicode_, iteritems, check_output
 
 
 def rb_escape(s):
@@ -341,74 +331,6 @@ def vagrant_setup_start(args):
         if retcode != 0:
             logging.critical("vagrant up failed with code %d", retcode)
             sys.exit(1)
-
-
-def find_ssh_executable(name='ssh'):
-    exts = os.environ.get('PATHEXT', '').split(os.pathsep)
-    dirs = list(os.environ.get('PATH', '').split(os.pathsep))
-    par, join = os.path.dirname, os.path.join
-    # executable might be bin/python or ReproUnzip\python
-    # or ReproUnzip\Python27\python or ReproUnzip\Python27\Scripts\something
-    loc = par(sys.executable)
-    local_dirs = []
-    for i in irange(3):
-        local_dirs.extend([loc, join(loc, 'ssh')])
-        loc = par(loc)
-    for pathdir in local_dirs + dirs:
-        for ext in exts:
-            fullpath = os.path.join(pathdir, name + ext)
-            if os.path.isfile(fullpath):
-                return fullpath
-    return None
-
-
-def run_interactive(ssh_info, interactive, cmds, request_pty):
-    if interactive:
-        ssh_exe = find_ssh_executable()
-    else:
-        ssh_exe = None
-
-    if interactive and ssh_exe:
-        record_usage(vagrant_ssh='ssh')
-        return subprocess.call(
-                [ssh_exe,
-                 '-t' if request_pty else '-T',  # Force allocation of PTY
-                 '-o', 'StrictHostKeyChecking=no',  # Silently accept host keys
-                 '-o', 'UserKnownHostsFile=/dev/null',  # Don't store host keys
-                 '-i', ssh_info['key_filename'],
-                 '-p', '%d' % ssh_info['port'],
-                 '%s@%s' % (ssh_info['username'],
-                            ssh_info['hostname']),
-                 cmds])
-    else:
-        record_usage(vagrant_ssh='interactive' if interactive else 'simple')
-        # Connects to the machine
-        ssh = paramiko.SSHClient()
-        ssh.set_missing_host_key_policy(IgnoreMissingKey())
-        ssh.connect(**ssh_info)
-
-        chan = ssh.get_transport().open_session()
-        if request_pty:
-            chan.get_pty()
-
-        # Execute command
-        logging.info("Connected via SSH, running command...")
-        chan.exec_command(cmds)
-
-        # Get output
-        if interactive:
-            interactive_shell(chan)
-        else:
-            chan.shutdown_write()
-            while True:
-                data = chan.recv(1024)
-                if len(data) == 0:
-                    break
-                sys.stdout.buffer.write(data)
-                sys.stdout.flush()
-        retcode = chan.recv_exit_status()
-        ssh.close()
-        return retcode
 
 
 @target_must_exist
